+++ changelog.md
@@ -3,6 +3,7 @@
 - Added Preview Banner button on the submission popup
 - Added Refund Button on the pending banners in your banners popup
 - Added "Limited Banners" filters
+- Added "Info" button on the banner cells.
 
 ## v0.1.3-beta.1